Port Second PR (Makes Verb Predicted)
