+++ Content.Server/_RMC/Examine/RMCSetPoseSystem.cs
@@ -1,49 +1,35 @@
 using Content.Server.Administration;
 using Content.Shared._RMC.Examine.Pose;
-using Content.Shared.Actions;
-using Content.Shared.Verbs;
+using Content.Shared.Administration.Logs;
+using Content.Shared.Database;
 using Robust.Shared.Player;
-using Robust.Shared.Utility;
 
 namespace Content.Server._RMC.Examine;
 
 public sealed class RMCSetPoseSystem : SharedRMCSetPoseSystem
 {
+    [Dependency] private readonly ISharedAdminLogManager _adminLog = default!;
     [Dependency] private readonly QuickDialogSystem _quickDialog = default!;
 
-    public override void Initialize()
+    protected override void SetPose(Entity<RMCSetPoseComponent> ent)
     {
-        base.Initialize();
-
-        SubscribeLocalEvent<RMCSetPoseComponent, GetVerbsEvent<Verb>>(OnSetPoseGetVerbs);
-    }
-
-    private void OnSetPoseGetVerbs(Entity<RMCSetPoseComponent> ent, ref GetVerbsEvent<Verb> args)
-    {
-        if (!args.CanInteract)
-            return;
-
-        if (args.User != args.Target)
-            return;
+        base.SetPose(ent);
 
         if (!TryComp<ActorComponent>(ent, out var actor))
             return;
 
         var setPosePrompt = Loc.GetString("rmc-set-pose-dialog", ("ent", ent));
-
-        Verb verb = new()
-        {
-            Text = Loc.GetString("rmc-set-pose-title"),
-            Icon = new SpriteSpecifier.Texture(new("/Textures/Interface/character.svg.192dpi.png")),
-            Priority = -5,
-            Act = () => _quickDialog.OpenDialog(actor.PlayerSession, Loc.GetString("rmc-set-pose-title"), setPosePrompt,
+        _quickDialog.OpenDialog(actor.PlayerSession,
+            Loc.GetString("rmc-set-pose-title"),
+            setPosePrompt,
             (string pose) =>
             {
+                if (pose.Length > 1000)
+                    pose = pose[..999];
+                _adminLog.Add(LogType.RMCSetPose, $"{ToPrettyString(ent)} set their pose to {pose}");
                 ent.Comp.Pose = pose;
                 Dirty(ent);
-            })
-        };
-
-        args.Verbs.Add(verb);
+            }
+        );
     }
 }
