+++ Content.Shared/_RMC/Examine/Pose/SharedRMCSetPoseSystem.cs
@@ -1,19 +1,38 @@
-using Content.Shared.Actions;
 using Content.Shared.Examine;
 using Content.Shared.Mobs;
+using Content.Shared.Verbs;
+using Robust.Shared.Utility;
 
 namespace Content.Shared._RMC.Examine.Pose;
 
 public abstract class SharedRMCSetPoseSystem : EntitySystem
 {
-    [Dependency] private readonly SharedActionsSystem _actions = default!;
-
     public override void Initialize()
     {
         base.Initialize();
 
+        SubscribeLocalEvent<RMCSetPoseComponent, GetVerbsEvent<Verb>>(OnSetPoseGetVerbs);
         SubscribeLocalEvent<RMCSetPoseComponent, ExaminedEvent>(OnExamine);
         SubscribeLocalEvent<RMCSetPoseComponent, MobStateChangedEvent>(OnMobStateChanged);
+    }
+
+    private void OnSetPoseGetVerbs(Entity<RMCSetPoseComponent> ent, ref GetVerbsEvent<Verb> args)
+    {
+        if (!args.CanInteract)
+            return;
+
+        if (args.User != args.Target)
+            return;
+
+        Verb verb = new()
+        {
+            Text = Loc.GetString("rmc-set-pose-title"),
+            Icon = new SpriteSpecifier.Texture(new ResPath("/Textures/Interface/character.svg.192dpi.png")),
+            Priority = -5,
+            Act = () => SetPose(ent),
+        };
+
+        args.Verbs.Add(verb);
     }
 
     private void OnExamine(Entity<RMCSetPoseComponent> ent, ref ExaminedEvent args)
@@ -37,5 +56,9 @@ public abstract class SharedRMCSetPoseSystem : EntitySystem
 
         ent.Comp.Pose = string.Empty; // reset the pose on death/crit
         Dirty(ent);
+    }
+
+    protected virtual void SetPose(Entity<RMCSetPoseComponent> ent)
+    {
     }
 }
